Only allow document api protocol version 5.115 or newer.
